大幅更新：画像のディレクトリ構成を修正

+++ app/Model/CrudBase.php
@@ -3,8 +3,8 @@
 /**
  * CrudBaseのロジッククラス
  * 
- * @version 2.3.0
- * @date 2016-1-21 | 2018-10-8
+ * @version 2.3.1
+ * @date 2016-1-21 | 2019-2-17
  * @history
  * 2018-10-8 v2.2.0 アップロードファイル関連の大幅修正
  * 2018-10-3 v2.2.0 アップロードファイルの抹消処理を追加
@@ -111,7 +111,7 @@ class CrudBase{
 	 */
 	public function guessDatetimeInfo($str,$option=array()){
 	    
-	    App::uses('DatetimeGuess','Vendor/Wacg');
+	    App::uses('DatetimeGuess','Vendor/CrudBase');
 	    if(empty($this->DatetimeGuess)) $this->DatetimeGuess = new DatetimeGuess();
 	    
 	    $info = $this->DatetimeGuess->guessDatetimeInfo($str,$option);
@@ -479,10 +479,8 @@ class CrudBase{
 	 * @param int $id
 	 * @param string $fn_field_strs ファイルフィールド群文字列（複数ある場合はコンマで連結）
 	 * @param array $ent エンティティ
-	 * @param string $dp_tmpl ディレクトリパス・テンプレート
-	 * @param string $viaDpFnMap 中継パスマッピング
 	 */
-	public function eliminateFiles(&$model,$id,$fn_field_strs, &$ent, $dp_tmpl, $viaDpFnMap){
+	public function eliminateFiles(&$model,$id,$fn_field_strs, &$ent){
 
 		// モデルと紐づいているテーブルからidに紐づくレコードを取得する。
 		$tbl_name = $model->useTable;
@@ -508,13 +506,11 @@ class CrudBase{
 			
 			// ▼削除データにファイルパスリストをセットする。
 			$fps = array();
-			
-			// ▼ ファイルパスを組み立てる
-			$fps['orig'] = $this->buildFp('orig', $fn_field, $ent, $dp_tmpl, $viaDpFnMap);
-			$fps['mid'] = $this->buildFp('mid', $fn_field, $ent, $dp_tmpl, $viaDpFnMap);
-			$fps['thum'] = $this->buildFp('thum', $fn_field, $ent, $dp_tmpl, $viaDpFnMap);
-			
 
+			// ▼ ファイルパスを組み立てる
+			$fps['orig'] = $ent[$fn_field];
+			$fps['mid'] = str_replace('/orig/', '/mid/', $ent[$fn_field]);
+			$fps['thum'] = str_replace('/orig/', '/thum/', $ent[$fn_field]);
 			$delEnt['fps'] = $fps;
 			
 			// ▼削除するファイル名が他のレコードで使われていないなら抹消フラグをONにする。使われているならfalseにする。
@@ -537,44 +533,16 @@ class CrudBase{
 			// ▼ファイルパスリストをループし、ファイルパスに紐づくファイルを削除する。
 			$fps = $delEnt['fps'];
 			foreach($fps as $fp){
-				unlink($fp);
+				if(file_exists($fp)){
+					unlink($fp);
+					
+				}
 			}
 			 
 		}
 
 	}
-	
-	/**
-	 * ファイルパスを組み立てる
-	 * @param string $dn ディレクトリ名
-	 * @param string $fn_field フィールド
-	 * @param array $ent データのエンティティ
-	 * @param string $dp_tmpl ディレクトリパス・テンプレート
-	 * @param array $viaDpFnMap 中継パスマッピング
-	 * @return string ファイルパス
-	 */
-	private function buildFp($dn, $fn_field, &$ent, $dp_tmpl, &$viaDpFnMap){
 
-		$fp = $dp_tmpl . $ent[$fn_field];
-		
-		$fp = str_replace('%field', $fn_field, $fp);
-		$fp = str_replace('%dn', $dn, $fp);
-		
-		// 経由ディレクトリパス部分を置換
-		$via_dp = ''; // 経由ディレクトリパス
-		if(!empty($viaDpFnMap[$fn_field])){
-			$via_dp_field = $viaDpFnMap[$fn_field];
-			$via_dp = $ent[$via_dp_field];
-		}
-		$fp = str_replace('%via_dp' , $via_dp , $fp );
-		$fp = str_replace('//' , '/' , $fp );
-		
-		
-		return $fp;
-	}
-	
-	
-	
 	
 
 }
